Check for config before accessing it
When a roam command is received in wifi connected state,
any existing wifi configuration for the current user is
searched based on the network ID .
This search can result in null i.e., no valid configuration
exists for the desired network ID.

The log misses this NULL check for config while accessing it.
Hence the check before the access.

CRs-Fixed: 2005496
Change-Id: I009579469a34d62af7270040dda88badf12eb2bf
Signed-off-by: Alex Naidis <alex.naidis@linux.com>